pyproject.toml
[tool.poetry]
name = "zezo"
version = "0.1.0"
description = "Um projeto muito brega para fazer você sofrer."
authors = ["felipebastos <felipebastos@ifce.edu.br>"]
readme = "README.md"
packages = [{ include = "zezo", from = "src" }]

[tool.taskipy.tasks]
run = { cmd = "python src/manage.py runserver", help = "Run the server" }
format = { cmd = 'black src', help = 'Formata o código' }
imports = { cmd = 'isort src', help = 'Organiza os imports do projeto' }

[tool.isort]
profile = "black"

[tool.black]
include = '\.pyi?$'


[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.3"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
pylint = "^3.0.1"
pre-commit = "^3.4.0"
taskipy = "^1.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
